b2c_invoice_back-main/requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-Cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
certifi
pydantic>=2.0.0,<3.0.0

# Google Cloud Document AI
google-cloud-documentai>=2.20.0

# LLM extraction (GPT-4o-mini)
openai>=1.12.0

# B2C Auth
bcrypt>=4.0.0
requests>=2.31.0

# Microsoft SSO
msal>=1.24.0

# Stripe payments
stripe>=7.0.0

# Excel file reading
openpyxl>=3.1.0
xlrd>=2.0.0

# Production WSGI server
gunicorn>=21.2.0
